+++ template_java/src/main/java/cs451/Main.java
@@ -88,10 +88,10 @@ public class Main {
         configArray = config.toArray(configArray);
 
         // Brodcast setup
-        // broadcaster = new LCBroadcast(parser.hosts(), parser.myId(), nb_msg,
+        // broadcaster = new LCBroadcast(parser.hosts(), (byte) parser.myId(), nb_msg,
         // configArray);
-        // broadcaster = new URBBroadcast(parser.hosts(), parser.myId(), nb_msg);
-        broadcaster = new FIFOBroadcast(parser.hosts(), parser.myId(), nb_msg);
+        // broadcaster = new URBBroadcast(parser.hosts(), (byte) parser.myId(), nb_msg);
+        broadcaster = new FIFOBroadcast(parser.hosts(), (byte) parser.myId(), nb_msg);
         // logfile setup
         path = Paths.get(parser.output());
